@@ -680,14 +706,14 @@
 		"@types/fs-extra": "^9.0.12",
 		"@types/glob": "^7.1.4",
 		"@types/mocha": "^9.0.0",
-		"@types/node": "^16.7.5",
-		"@types/node-fetch": "^2.5.12",
+		"@types/node": "^16.9.1",
+		"@types/node-fetch": "^3.0.3",
 		"@types/power-assert": "^1.5.4",
 		"@types/uuid": "^8.3.1",
-		"@types/vscode": "^1.59.0",
-		"@typescript-eslint/eslint-plugin": "^4.29.3",
-		"@typescript-eslint/parser": "^4.29.3",
-		"electron": "^13.2.3",
+		"@types/vscode": "^1.60.0",
+		"@typescript-eslint/eslint-plugin": "^4.31.0",
+		"@typescript-eslint/parser": "^4.31.0",
+		"electron": "^14.0.0",
 		"eslint": "^7.32.0",
 		"file-loader": "^6.2.0",
 		"glob": "^7.1.7",
@@ -695,10 +721,10 @@
 		"power-assert": "^1.6.1",
 		"ts-loader": "^9.2.5",
 		"ts-node": "^10.2.1",
-		"typescript": "^4.4.2",
+		"typescript": "^4.4.3",
 		"vscode-debugadapter-testsupport": "^1.49.0",
 		"vscode-test": "^1.6.1",
-		"webpack": "^5.51.1",
+		"webpack": "^5.52.1",
 		"webpack-cli": "^4.8.0"
 	},
 	"optionalDependencies": {
