@@ -1,11 +1,18 @@
 "use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
+/* ***** BEGIN LICENSE BLOCK *****
+    Copyright (c) 2018-2021 Famibee (famibee.blog38.fc2.com)
+
+    This software is released under the MIT License.
+    http://opensource.org/licenses/mit-license.php
+** ***** END LICENSE BLOCK ***** */
+exports.__esModule = true;
 exports.getFn = exports.chkBoolean = exports.replaceFile = exports.foldProc = exports.treeProc = exports.regNoUseSysPath = exports.statBreak = exports.is_mac = exports.is_win = exports.oIcon = exports.setCtx4 = exports.getNonce = exports.docsel = exports.REG_SCRIPT = exports.uint = exports.int = void 0;
-const vscode_1 = require("vscode");
+var vscode_1 = require("vscode");
+// =============== Global
 function int(o) { return parseInt(String(o), 10); }
 exports.int = int;
 function uint(o) {
-    const v = parseInt(String(o), 10);
+    var v = parseInt(String(o), 10);
     return v < 0 ? -v : v;
 }
 exports.uint = uint;
@@ -14,9 +21,9 @@ exports.docsel = { scheme: 'file', language: 'skynovel' };
 ;
 ;
 function getNonce() {
-    let text = '';
-    const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    for (let i = 0; i < 32; i++) {
+    var text = '';
+    var possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    for (var i = 0; i < 32; i++) {
         text += possible.charAt(Math.floor(Math.random() * possible.length));
     }
     return text;
@@ -24,34 +31,36 @@ function getNonce() {
 exports.getNonce = getNonce;
 function setCtx4(ctx0) { extPath = ctx0.extensionPath; }
 exports.setCtx4 = setCtx4;
-let extPath = '';
+var extPath = '';
 function oIcon(name) {
     return {
-        light: `${extPath}/res/light/${name}.svg`,
-        dark: `${extPath}/res/dark/${name}.svg`,
+        light: extPath + "/res/light/" + name + ".svg",
+        dark: extPath + "/res/dark/" + name + ".svg"
     };
 }
 exports.oIcon = oIcon;
 ;
 exports.is_win = process.platform === 'win32';
 exports.is_mac = process.platform === 'darwin';
-exports.statBreak = exports.is_mac ? () => '&&'
-    : exports.is_win ? () => {
-        const chkShell = String(vscode_1.workspace.getConfiguration('terminal.integrated.shell').get('windows')).slice(-7);
+//const is_linux = process.platform === 'linux';
+exports.statBreak = exports.is_mac ? function () { return '&&'; }
+    : exports.is_win ? function () {
+        var chkShell = String(vscode_1.workspace.getConfiguration('terminal.integrated.shell').get('windows')).slice(-7);
         return (chkShell === 'cmd.exe') ? '&' : ';';
     }
-        : () => ';';
-const m_fs = require("fs-extra");
-const path_1 = require("path");
-const regNoUseSysFile = /^(\..+|.+\.(db|ini|git)|_notes|Icon\r)$/;
+        : function () { return ';'; };
+// 階層フォルダ逐次処理
+var m_fs = require("fs-extra");
+var path_1 = require("path");
+var regNoUseSysFile = /^(\..+|.+\.(db|ini|git)|_notes|Icon\r)$/;
 exports.regNoUseSysPath = /\/(\..+|.+\.(db|ini|git)|_notes|Icon\r)$/;
 function treeProc(wd, fnc) {
-    m_fs.readdirSync(wd, { withFileTypes: true }).forEach((d) => {
+    m_fs.readdirSync(wd, { withFileTypes: true }).forEach(function (d) {
         regNoUseSysFile.lastIndex = 0;
-        const nm = String(d.name).normalize('NFC');
+        var nm = String(d.name).normalize('NFC');
         if (regNoUseSysFile.test(nm))
             return;
-        const url = path_1.resolve(wd, nm);
+        var url = (0, path_1.resolve)(wd, nm);
         if (d.isDirectory()) {
             treeProc(url, fnc);
             return;
@@ -61,43 +70,48 @@ function treeProc(wd, fnc) {
 }
 exports.treeProc = treeProc;
 function foldProc(wd, fnc, fncFld) {
-    m_fs.readdirSync(wd, { withFileTypes: true }).forEach((d) => {
+    m_fs.readdirSync(wd, { withFileTypes: true }).forEach(function (d) {
         regNoUseSysFile.lastIndex = 0;
-        const nm = String(d.name).normalize('NFC');
+        var nm = String(d.name).normalize('NFC');
         if (regNoUseSysFile.test(nm))
             return;
         if (d.isDirectory()) {
             fncFld(nm);
             return;
         }
-        const url = path_1.resolve(wd, nm);
+        var url = (0, path_1.resolve)(wd, nm);
         fnc(url, nm);
     });
 }
 exports.foldProc = foldProc;
-function replaceFile(src, r, rep, dest = src) {
+function replaceFile(src, r, rep, dest) {
+    if (dest === void 0) { dest = src; }
     try {
         if (!m_fs.existsSync(src))
             return;
-        const txt = m_fs.readFileSync(src, { encoding: 'utf8' });
-        const ret = String(txt.replace(r, rep));
+        var txt = m_fs.readFileSync(src, { encoding: 'utf8' });
+        var ret = String(txt.replace(r, rep));
         m_fs.ensureFileSync(dest);
         if (txt !== ret)
             m_fs.writeFileSync(dest, ret);
     }
     catch (err) {
-        console.error(`replaceFile src:${src} ${err}`);
+        console.error("replaceFile src:" + src + " " + err);
     }
 }
 exports.replaceFile = replaceFile;
+/*export	function argChk_Boolean(hash: any, name: string, def: boolean): boolean {
+    if (! (name in hash)) return hash[name] = def;
+
+    return hash[name] = chkBoolean(hash[name]);
+}*/
 function chkBoolean(v) {
     if (v === null)
         return false;
-    const v2 = String(v);
+    var v2 = String(v);
     return (v2 === 'false') ? false : Boolean(v2);
 }
 exports.chkBoolean = chkBoolean;
-function getFn(path) { return path_1.basename(path, path_1.extname(path)); }
+function getFn(path) { return (0, path_1.basename)(path, (0, path_1.extname)(path)); }
 exports.getFn = getFn;
 ;
-//# sourceMappingURL=CmnLib.js.map
