@@ -11,7 +11,7 @@ import {TreeItem, TreeItemCollapsibleState, commands, ExtensionContext, ThemeIco
 
 export interface TREEITEM_CFG {
 	cmd		: string,
-	dbg?	: boolean,
+	exe?	: boolean,
 	icon	: string,
 	label	: string,
 	desc?	: string,
@@ -20,21 +20,26 @@ export interface TREEITEM_CFG {
 	forMac?		: boolean,
 }
 
-export class MyTreeItem extends TreeItem {
+export class PrjTreeItem extends TreeItem {
 	private		_children	: TreeItem[]	= [];
 
+	// TODO: registerCommand(の登録が複数プロジェクトで重複する
+
+	readonly	exe: boolean;
 	constructor(readonly cfg: TREEITEM_CFG, readonly dir: string, readonly ctx: ExtensionContext, readonly onClickTreeItemBtn: (ti: TreeItem, btn_nm: string, cfg: TREEITEM_CFG)=> void) {
 		super(is_win && cfg.forMac ?'' :cfg.label);
 
+		this.exe = Boolean(cfg.exe);
 		if (is_win && cfg.forMac) this.description = '（Windowsでは使えません）';
 		else {
 			this.description = cfg.desc ?? '';
 			if (cfg.cmd) {
 				const btn_nm = this.contextValue = 'skynovel.dev'+ cfg.cmd;
 				ctx.subscriptions.push(commands.registerCommand(btn_nm, ti=> onClickTreeItemBtn(ti, cfg.cmd, cfg)));
-				if (cfg.dbg) {
-					const btn_nm2 = btn_nm +'Dbg';
-					ctx.subscriptions.push(commands.registerCommand(btn_nm2, ti=> onClickTreeItemBtn(ti, cfg.cmd +'Dbg', cfg)));
+				if (cfg.exe) {
+					ctx.subscriptions.push(commands.registerCommand(btn_nm +'Dbg', ti=> onClickTreeItemBtn(ti, cfg.cmd +'Dbg', cfg)));
+
+					ctx.subscriptions.push(commands.registerCommand(btn_nm +'Stop', ti=> onClickTreeItemBtn(ti, cfg.cmd +'Stop', cfg)));
 				}
 			}
 		}
@@ -42,7 +47,7 @@ export class MyTreeItem extends TreeItem {
 		if (cfg.children) {
 			this.iconPath = ThemeIcon.Folder;
 			this.collapsibleState = TreeItemCollapsibleState.Collapsed;
-			this._children = cfg.children.map(cCfg=> new MyTreeItem(cCfg, dir, ctx, onClickTreeItemBtn));
+			this._children = cfg.children.map(cCfg=> new PrjTreeItem(cCfg, dir, ctx, onClickTreeItemBtn));
 		}
 		else {
 			this.iconPath = oIcon(cfg.icon);
